Implement solution to margin/padding ex 02
This implements my personal solution to margins
and padding exercise 02, using my best judgment
about how best to arrange things. Getting the
first item in the card the right distance from
the top and center-aligning the text work in
my solution, but I'm not sure if I did it the
ideal way.